AKK band L added, multi-band support refined

+++ libraries/AP_VideoTX/AP_SmartAudio.h
@@ -235,8 +235,8 @@ private:
     void set_operation_mode(uint8_t mode);
     // change the frequency
     void set_frequency(uint16_t frequency, bool isPitModeFreq);
-    // change the channel
-    void set_channel(uint8_t chan);
+    // change the channel that already includes its band
+    void set_channel(uint16_t chan);
     // get the pitmode frequency
     void request_pit_mode_frequency();
     // set the power
